Fix cache saving problem

+++ src/actions/Actions.js
@@ -145,23 +145,19 @@ export let addToWatchlist = (userId, sessionId, filmId, isAdding) => async (disp
     }
 }
 
-export let rateFilm = (sessionId, filmId, rating) => async (dispatch) => {
+export let changeFilmRating = (sessionId, filmId, rating) => async (dispatch) => {
     try {
         dispatch({type: 'SET_FILM_BUTTONS_LOADING'});
-        let response = await service.rateFilm(sessionId, filmId, rating);
+
+        let response;
+        if (rating) {
+            response = await service.rateFilm(sessionId, filmId, rating);
+        } else {
+            response = await service.deleteFilmRating(sessionId, filmId);
+        }
+
         if (response.status_code !== 1 && response.status_code !== 12 && response.status_code !== 13) throw new Error('Что-то пошло не так, попробуйте позже');
         dispatch({type: 'SET_FILM_RATING', payload: rating});
-    } catch (err) {
-        dispatch({type: 'SET_FILM_STATES_ERROR', payload: err.message});
-    }
-}
-
-export let deleteFilmRating = (sessionId, filmId) => async (dispatch) => {
-    try {
-        dispatch({type: 'SET_FILM_BUTTONS_LOADING'});
-        let response = await service.deleteFilmRating(sessionId, filmId);
-        if (response.status_code !== 1 && response.status_code !== 12 && response.status_code !== 13) throw new Error('Что-то пошло не так, попробуйте позже');
-        dispatch({type: 'SET_FILM_RATING', payload: false});
     } catch (err) {
         dispatch({type: 'SET_FILM_STATES_ERROR', payload: err.message});
     }
